Bot/requirements.txt
gradio
huggingface_hub>=0.25.2
python-dotenv
pydantic==2.10.6

